offb: Fix setting of the pseudo-palette for >8bpp
commit 1bb0b7d21584b3f878e2bc880db62351ddee5185 upstream.

When using a >8bpp framebuffer, offb advertises truecolor, not
directcolor,
and doesn't touch the color map even if it has a corresponding access
method
for the real hardware.

Thus it needs to set the pseudo-palette with all 3 components of the
color,
like other truecolor framebuffers, not with copies of the color index
like
a directcolor framebuffer would do.

This went unnoticed for a long time because it's pretty hard to get offb
to kick in with anything but 8bpp (old BootX under MacOS will do that
and
qemu does it).

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>

Conflicts:
	drivers/video/offb.c